moved things from main.java to TestDriver.java
